adk-2.0/pyproject.toml
[project]
name = "adk-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=2.0.0a2",
]
